+++ apps/api-register/src/pages/api/altcha/challenge.ts
@@ -1,10 +1,14 @@
 import { ALTCHA_HMAC_KEY } from "astro:env/server";
 import { createChallenge } from "altcha-lib";
+import { deriveKey } from "altcha-lib/algorithms/pbkdf2";
 import type { APIRoute } from "astro";
 
 export const GET: APIRoute = async () => {
   const challenge = await createChallenge({
-    hmacKey: ALTCHA_HMAC_KEY,
+    algorithm: "PBKDF2/SHA-256",
+    cost: 5000,
+    deriveKey,
+    hmacSignatureSecret: ALTCHA_HMAC_KEY,
   });
 
   return new Response(JSON.stringify(challenge), {