+++ packages/layouts/src/helpers/getPagination.test.ts
@@ -130,10 +130,15 @@ test("parseHeaders() should return default values when headers are missing", () 
   const headers = new Headers();
 
   expect(parseHeaders(headers)).toEqual({
-    currentPage: 0,
+    currentPage: 1,
     perPage: 0,
     totalCount: 0,
-    totalPages: 0,
+    totalPages: 1,
+    self: {
+      page: "1",
+      rel: "self",
+      url: "",
+    },
   });
 });
 
@@ -147,7 +152,7 @@ test("parseHeaders() should return default values when some headers are missing,
 
   expect(parseHeaders(headers)).toEqual({
     currentPage: 0,
-    perPage: 0,
+    perPage: 26,
     totalCount: 26,
     totalPages: 0,
     first: {
@@ -155,18 +160,24 @@ test("parseHeaders() should return default values when some headers are missing,
       _per_page: "10",
       rel: "first",
       url: "http://my-json-server.typicode.com/developer-overheid-nl/dingen-api/things?_page=1&_per_page=10",
+      page: undefined,
+      perPage: undefined,
     },
     last: {
       _page: "3",
       _per_page: "10",
       rel: "last",
       url: "http://my-json-server.typicode.com/developer-overheid-nl/dingen-api/things?_page=3&_per_page=10",
+      page: undefined,
+      perPage: undefined,
     },
     next: {
       _page: "2",
       _per_page: "10",
       rel: "next",
       url: "http://my-json-server.typicode.com/developer-overheid-nl/dingen-api/things?_page=2&_per_page=10",
+      page: undefined,
+      perPage: undefined,
     },
   });
 });
@@ -213,6 +224,24 @@ test("parseHeaders() should return default values when some headers are missing,
   });
 });
 
+test("parseHeaders() should return correct values when link is missing", () => {
+  const headers = new Headers();
+  headers.set("x-total-count", "6");
+  headers.set("link", "");
+
+  expect(parseHeaders(headers)).toEqual({
+    currentPage: 1,
+    perPage: 6,
+    totalCount: 6,
+    totalPages: 1,
+    self: {
+      page: "1",
+      rel: "self",
+      url: "",
+    },
+  });
+});
+
 test("getPagination() should return correct pagination links and range, first page", () => {
   const pagination = {
     currentPage: 1,
@@ -240,12 +269,12 @@ test("getPagination() should return correct pagination links and range, first pa
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url)).toEqual({
     links: [
       {
-        href: "https://api.example.com/1",
+        href: "https://api.example.com/items/1",
         label: 1,
         range: [1, 10],
         current: true,
@@ -257,11 +286,11 @@ test("getPagination() should return correct pagination links and range, first pa
     current: 1,
     ellipsisAfter: true,
     next: {
-      href: "https://api.example.com/2",
+      href: "https://api.example.com/items/2",
       label: 2,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -299,34 +328,34 @@ test("getPagination() should return correct pagination links and range, page 2",
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url)).toEqual({
     links: [
       {
-        href: "https://api.example.com/2",
+        href: "https://api.example.com/items/2",
         label: 2,
         range: [11, 20],
         current: true,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     ellipsisBefore: false,
     current: 2,
     ellipsisAfter: true,
     next: {
-      href: "https://api.example.com/3",
+      href: "https://api.example.com/items/3",
       label: 3,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -364,34 +393,34 @@ test("getPagination() should return correct pagination links and range, page 3",
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url)).toEqual({
     links: [
       {
-        href: "https://api.example.com/3",
+        href: "https://api.example.com/items/3",
         label: 3,
         range: [21, 30],
         current: true,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/2",
+      href: "https://api.example.com/items/2",
       label: 2,
     },
     ellipsisBefore: true,
     current: 3,
     ellipsisAfter: true,
     next: {
-      href: "https://api.example.com/4",
+      href: "https://api.example.com/items/4",
       label: 4,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -429,34 +458,34 @@ test("getPagination() should return correct pagination links and range, last pag
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url)).toEqual({
     links: [
       {
-        href: "https://api.example.com/41",
+        href: "https://api.example.com/items/41",
         label: 41,
         range: [401, 410],
         current: true,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/40",
+      href: "https://api.example.com/items/40",
       label: 40,
     },
     ellipsisBefore: true,
     current: 41,
     ellipsisAfter: false,
     next: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -489,23 +518,23 @@ test("getPagination() should return correct pagination links and range, last pag
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url)).toEqual({
     links: [
       {
-        href: "https://api.example.com/42",
+        href: "https://api.example.com/items/42",
         label: 42,
         range: [411, 420],
         current: true,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/41",
+      href: "https://api.example.com/items/41",
       label: 41,
     },
     ellipsisBefore: true,
@@ -548,44 +577,44 @@ test("getPagination() should return correct pagination links and range, page 15;
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url, 1)).toEqual({
     links: [
       {
-        href: "https://api.example.com/14",
+        href: "https://api.example.com/items/14",
         label: 14,
         current: false,
       },
       {
-        href: "https://api.example.com/15",
+        href: "https://api.example.com/items/15",
         label: 15,
         range: [141, 150],
         current: true,
       },
       {
-        href: "https://api.example.com/16",
+        href: "https://api.example.com/items/16",
         label: 16,
         current: false,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/14",
+      href: "https://api.example.com/items/14",
       label: 14,
     },
     ellipsisBefore: true,
     current: 15,
     ellipsisAfter: true,
     next: {
-      href: "https://api.example.com/16",
+      href: "https://api.example.com/items/16",
       label: 16,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -623,54 +652,54 @@ test("getPagination() should return correct pagination links and range, page 15;
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url, 2)).toEqual({
     links: [
       {
-        href: "https://api.example.com/13",
+        href: "https://api.example.com/items/13",
         label: 13,
         current: false,
       },
       {
-        href: "https://api.example.com/14",
+        href: "https://api.example.com/items/14",
         label: 14,
         current: false,
       },
       {
-        href: "https://api.example.com/15",
+        href: "https://api.example.com/items/15",
         label: 15,
         range: [141, 150],
         current: true,
       },
       {
-        href: "https://api.example.com/16",
+        href: "https://api.example.com/items/16",
         label: 16,
         current: false,
       },
       {
-        href: "https://api.example.com/17",
+        href: "https://api.example.com/items/17",
         label: 17,
         current: false,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/14",
+      href: "https://api.example.com/items/14",
       label: 14,
     },
     ellipsisBefore: true,
     current: 15,
     ellipsisAfter: true,
     next: {
-      href: "https://api.example.com/16",
+      href: "https://api.example.com/items/16",
       label: 16,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -708,49 +737,49 @@ test("getPagination() should return correct pagination links and range, page 40;
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url, 2)).toEqual({
     links: [
       {
-        href: "https://api.example.com/38",
+        href: "https://api.example.com/items/38",
         label: 38,
         current: false,
       },
       {
-        href: "https://api.example.com/39",
+        href: "https://api.example.com/items/39",
         label: 39,
         current: false,
       },
       {
-        href: "https://api.example.com/40",
+        href: "https://api.example.com/items/40",
         label: 40,
         range: [391, 400],
         current: true,
       },
       {
-        href: "https://api.example.com/41",
+        href: "https://api.example.com/items/41",
         label: 41,
         current: false,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/39",
+      href: "https://api.example.com/items/39",
       label: 39,
     },
     ellipsisBefore: true,
     current: 40,
     ellipsisAfter: false,
     next: {
-      href: "https://api.example.com/41",
+      href: "https://api.example.com/items/41",
       label: 41,
     },
     last: {
-      href: "https://api.example.com/42",
+      href: "https://api.example.com/items/42",
       label: 42,
     },
   });
@@ -788,33 +817,33 @@ test("getPagination() should return correct pagination links and range, last pag
       url: "https://api.example.com/items?page=42",
     },
   };
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
 
   expect(getPagination(pagination, url, 2)).toEqual({
     links: [
       {
-        href: "https://api.example.com/40",
+        href: "https://api.example.com/items/40",
         label: 40,
         current: false,
       },
       {
-        href: "https://api.example.com/41",
+        href: "https://api.example.com/items/41",
         label: 41,
         current: false,
       },
       {
-        href: "https://api.example.com/42",
+        href: "https://api.example.com/items/42",
         label: 42,
         range: [411, 420],
         current: true,
       },
     ],
     first: {
-      href: "https://api.example.com/1",
+      href: "https://api.example.com/items/1",
       label: 1,
     },
     prev: {
-      href: "https://api.example.com/41",
+      href: "https://api.example.com/items/41",
       label: 41,
     },
     ellipsisBefore: true,
@@ -826,15 +855,123 @@ test("getPagination() should return correct pagination links and range, last pag
 });
 
 test("getPagination() should return empty links when pagination is null", () => {
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
   expect(getPagination(null, url)).toEqual({
     links: [],
   });
 });
 
 test("getPagination() should return empty links when pagination is undefined", () => {
-  const url = new URL("https://api.example.com/items");
+  const url = new URL("https://api.example.com/items/1");
   expect(getPagination(undefined, url)).toEqual({
     links: [],
+  });
+});
+
+test("getPagination() should return correct pagination links when currentPage is 1 and totalPages is 1", () => {
+  const pagination = {
+    currentPage: 1,
+    perPage: 6,
+    totalCount: 6,
+    totalPages: 1,
+    self: {
+      page: "1",
+      rel: "self",
+      url: "",
+    },
+  };
+  const url = new URL("https://api.example.com/items/1");
+
+  expect(getPagination(pagination, url)).toEqual({
+    links: [
+      {
+        href: "https://api.example.com/items/1",
+        label: 1,
+        range: [1, 6],
+        current: true,
+      },
+    ],
+    first: false,
+    prev: false,
+    ellipsisBefore: false,
+    current: 1,
+    ellipsisAfter: false,
+    next: false,
+    last: false,
+  });
+});
+
+test("getPagination() should return correct pagination links when they have a ampersand in the url", () => {
+  const pagination = {
+    currentPage: 41,
+    perPage: 10,
+    totalCount: 420,
+    totalPages: 42,
+    self: {
+      page: "41",
+      rel: "self",
+      url: "https://api.example.com/items?page=41&perPage=10&filter=this+%26%20that",
+    },
+    prev: {
+      page: "40",
+      rel: "prev",
+      url: "https://api.example.com/items?page=40&perPage=10&filter=this+%26%20that",
+    },
+    first: {
+      page: "1",
+      rel: "first",
+      url: "https://api.example.com/items?page=1&perPage=10&filter=this+%26%20that",
+    },
+    next: {
+      page: "42",
+      rel: "next",
+      url: "https://api.example.com/items?page=42&perPage=10&filter=this+%26%20that",
+    },
+    last: {
+      page: "42",
+      rel: "last",
+      url: "https://api.example.com/items?page=42&perPage=10&filter=this+%26%20that",
+    },
+  };
+  const url = new URL("https://api.example.com/items/1?filter=this+%26%20that");
+
+  expect(getPagination(pagination, url, 2)).toEqual({
+    links: [
+      {
+        href: "https://api.example.com/items/39?filter=this+%26%20that",
+        label: 39,
+        current: false,
+      },
+      {
+        href: "https://api.example.com/items/40?filter=this+%26%20that",
+        label: 40,
+        current: false,
+      },
+      {
+        href: "https://api.example.com/items/41?filter=this+%26%20that",
+        label: 41,
+        range: [401, 410],
+        current: true,
+      },
+    ],
+    first: {
+      href: "https://api.example.com/items/1?filter=this+%26%20that",
+      label: 1,
+    },
+    prev: {
+      href: "https://api.example.com/items/40?filter=this+%26%20that",
+      label: 40,
+    },
+    ellipsisBefore: true,
+    current: 41,
+    ellipsisAfter: false,
+    next: {
+      href: "https://api.example.com/items/42?filter=this+%26%20that",
+      label: 42,
+    },
+    last: {
+      href: "https://api.example.com/items/42?filter=this+%26%20that",
+      label: 42,
+    },
   });
 });